Lagos_House_Price
This is my first Machine Learning Project. Let me provide a background. This a dataset from a real estate company in Lagos Nigeria. This dataset includes prices of houses in Lagos area, the facilities such as CCTV, apartment type, other facilities. I was able to generate an insight as to the price range by location, by facilities and thereafter using two ML algorithm DecisionTree and RandomForest.  This Dataset was trained with 30% and 42% Random State to give me a trend when reproduced.