Blossom Bank Online Machine Learning Fraud Detection
BACKGROUND
Blossom Bank also known as BB PLC is a multinational financial services group, that 
offers retail and investment banking, pension management, asset management, and 
payments services, headquartered in London, UK.
Problem:
Blossom Bank wants to build a Machine Learning model to predict online payment fraud.
I prepare the data for mining and EDA
I checked for missing values, renamed the columns for clearer understanding by the management
I also included Visuals of Transactions such as:
Types of transactions, 
Time of Transactions
Disparity between Fraudulent Transaction and Non Fraudulent
 The percentage of fraudulent Transactions which is about 0.1%
Using Univariate Bivariate and Multivariate Analysis with Pie Chart, Boxplot, Heatmap, Scattered plot and so on.
To prepare the dataset for modeling and machine learning algorithm, a few changes are require with the column with the fraudulent transaction.
I therefore, make sure to separate this column to know actual number of transactions that were fraudulent. Doing this give me an insight to a fact that there is an imbalance in the dataset. Actually, for us to get a reliable model, we need a balance dataset. I therefore used the actual number of the fraud transaction sample of 1142 to create another 1142 Non fraud transaction and joined it together. I checked for the statistical balance so that it does not deviate from the original dataset and it was accurate. I went ahead to removed the categorical columns by converting them to float using One-Hot Encoding. 
I set my target to isFrausd column and used two Machine Learning Models. Decision Tree and K-Neighbor
Decision Tree gave me 95% - 97% Accuracy, Precision and Recall compared to K-Neighbor with approximately 90%. 
My conclusion therefore is since, Decision tree gives us 95% prediction of online payment fraud, it is therefore the best model.